// tests/e2e/orm/plugin-infra/ext-query-args/schema.ts
//////////////////////////////////////////////////////////////////////////////////////////////
// DO NOT MODIFY THIS FILE                                                                  //
// This file is automatically generated by ZenStack CLI and should not be manually updated. //
//////////////////////////////////////////////////////////////////////////////////////////////

/* eslint-disable */

import { type SchemaDef, ExpressionUtils } from "@zenstackhq/orm/schema";
export class SchemaType implements SchemaDef {
    provider = {
        type: "sqlite"
    } as const;
    models = {
        User: {
            name: "User",
            fields: {
                id: {
                    name: "id",
                    type: "Int",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("autoincrement") }] }],
                    default: ExpressionUtils.call("autoincrement")
                },
                name: {
                    name: "name",
                    type: "String"
                }
            },
            idFields: ["id"],
            uniqueFields: {
                id: { type: "Int" }
            }
        }
    } as const;
    authType = "User" as const;
    plugins = {};
}
export const schema = new SchemaType();